add `IngestionIntensivity` chart
implement "Phase Visualization" which is a plot showing up where's the
peak on timeline and when ingestion will have it's strongest effects. I
do not count tears on this features and how trouble direct rendering
into terminal instead string can cause, anyway... as far my testing of
this feature was going - it seems to be mostly fine, will need a little
refactor in near time but for now it's completly fine. #522