add experimental `monitor` command for ingestion intensity visualization
introduces a new `monitor` command to the CLI which launches an interactive Terminal User Interface (TUI). This interface visualizes the effects of substance ingestion over time, aiding in substance effect tracking and monitoring. A comprehensive refactor was carried out within the documentation to explain the usage of this new feature, and minor updates were also made to `src/main.rs` to launch the TUI. Moreover, the application's database structure was updated to accommodate the needed information for this monitor, as shown in modifications to `src/tui` and `docs/composite.md`.